2011-10-11	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libcsupport/src/termios.c,
	libcsupport/include/rtems/termiostypes.h (rtems_termios_puts):
	Use size_t for buffer size.

+++ cpukit/libcsupport/include/rtems/termiostypes.h
@@ -164,7 +164,7 @@ void rtems_termios_rxirq_occured(struct rtems_termios_tty *tty);
  */
 void rtems_termios_puts (
   const void               *buf,
-  int                       len,
+  size_t                    len,
   struct rtems_termios_tty *tty
 );
